+++ examples/resources/vaultwarden_user_invite/import.sh
@@ -1,1 +1,0 @@
-terraform import vaultwarden_user_invite.example <id>