+++ examples/resources/vaultwarden_user_invite/resource.tf
@@ -1,3 +1,0 @@
-resource "vaultwarden_user_invite" "example" {
-  email = "foo@example.com"
-}